requirements.txt
openai==0.28
fastapi==0.110.0 
uvicorn==0.29.0 
python-dotenv==1.0.0
python-multipart==0.0.9